Moved instructions card collection delegates into its own file.
